Acerto no input do procedimento.

+++ SQLQueryProjProc.sql
@@ -82,7 +82,7 @@ IF OBJECT_ID ('SchemaLicitacao.procLicitarProd', 'P') IS NOT NULL
 	DROP proc SchemaLicitacao.procLicitarProd;
 GO
 Create proc SchemaLicitacao.procLicitarProd
-			(@userID int, @prodid int, @licitaval int)
+			(@userID int, @prodid int, @licitaval decimal)
 as
 BEGIN
 	DECLARE @msgErro varchar(500)
